moved completion functions:
